refactor(indexer): phase 1 — extract metrics HTTP server
Move `run_metrics_server` out of `indexer-httpd` into a new
`indexer-metrics` crate at `indexer/metrics/`. The metrics endpoint
just renders Prometheus scrapes from a `PrometheusHandle` — it has
no dependency on SQL state or indexer wiring, so it doesn't belong
in `indexer-httpd`.

Also delete an identical, dead copy of `run_metrics_server` from
`grug-httpd` (zero callers in the workspace) and drop the now-unused
`metrics-exporter-prometheus` dep from both `indexer-httpd` and
`grug-httpd`.

The new crate returns `Result<(), io::Error>` directly instead of
wrapping it in a custom error enum; the only fallible calls inside
are `bind()` and `run()`, both already `io::Error`.
