+++ indexer/metrics/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+authors       = { workspace = true }
+categories    = { workspace = true }
+documentation = { workspace = true }
+edition       = { workspace = true }
+license       = { workspace = true }
+name          = "indexer-metrics"
+repository    = { workspace = true }
+rust-version  = { workspace = true }
+version       = { workspace = true }
+
+[lib]
+path = "src/lib.rs"
+
+[features]
+tracing = ["dep:tracing"]
+
+[dependencies]
+actix-web                   = { workspace = true }
+actix-web-metrics           = { workspace = true }
+metrics-exporter-prometheus = { workspace = true }
+sentry-actix                = { workspace = true }
+tracing                     = { workspace = true, optional = true }
